remote/requirements.txt
beautifulsoup4
pwntools
requests
pycryptodome